Add WAbstractSlider and do some initial code for labeled sliders
